feat:  remove location confirmation

+++ js/app.controller.js
@@ -69,6 +69,8 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    if(!confirm("are you sure you want to delete thhis location?")) return
+
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
